Move add-record page to core/pages, create core module
