+++ src/app/app.module.ts
@@ -17,7 +17,7 @@ export function LanguageLoader(http: HttpClient) {
 
 import { LottieModule } from 'ngx-lottie';
 import player from 'lottie-web';
-import { AddRecordPageModule } from './add-record/add-record.module';
+import { CoreModule } from './core/core.module';
 
 // Note we need a separate function as it's required
 // by the AOT compiler.
@@ -38,7 +38,7 @@ export function playerFactory() {
       deps: [HttpClient]
       }
     }),
-    AddRecordPageModule,
+    CoreModule,
     LottieModule.forRoot({ player: playerFactory })
 ],
   providers: [
